+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
+Unreleased
+----------
+
+* Refactored `AbstractModel`
+
 2.1.0
 -----